+++ CONTRIBUTING.md
@@ -1,0 +1,65 @@
+# How to contribute?
+
+Thanks for thinking about contributing to this project! You can contribute in one of several ways:
+
+- [Find bugs!](#find-bugs-!)
+- [Fix bugs!](#fix-bugs-!)
+- [Add new features or improve an existing one!](#add-new-features-or-improve-an-existing-one!)
+- [Do you have questions?](#Do-you-have-questions?)
+
+## Find bugs!
+
+Yes! The easiest way to contribute is to make sure the existing code works as expected! So if you think you found a bug, please report it so we can act on it!
+
+Ensure the bug was not already reported by searching on GitHub Issues. This helps prevent duplication.
+
+If you're unable to find an open issue addressing the problem, open a new one. Be sure to include a title and clear description, as much relevant information as possible, and a code sample or an executable test case demonstrating the expected behavior that is not occurring.
+
+If possible, use the relevant bug report templates to create the issue. Simply copy the content of the appropriate template into a .rb file, make the necessary changes to demonstrate the issue, and paste the content into the issue description (some examples below):
+
+- Error reporting tool issues
+
+- Error classification issues
+
+- Robot visualization issues
+
+- Listener agent issues
+
+- Remote intervention issues
+
+## Fix bugs!
+
+- Help us keep the code bug free!
+
+- Please look through existing issues and choose one to work on. These have already been triaged by maintainers to be important things to fix and have the most impact on the project.
+
+- Alternatively, work on a patch for a bug you found as well.
+
+- Please refer to the branching model on [BRANCHING_MODEL.md](/BRANCHING_MODEL.md) file for best practices on how to use branches for this project.
+
+- Please refer to the ROS C++ styling guide that is based on Google C++ style guide as much as possible to keep the code readable!
+
+- Work on the patch.
+
+- Open a new GitHub pull request with the patch.
+
+- Ensure the pull request description clearly describes the problem and solution. Include the relevant issue number if applicable.
+
+## Add new features or improve an existing one!
+
+- Do not open an issue on GitHub. These are primarily intended for bug reports and fixes.
+
+- Just follow the branching model on [BRANCHING_MODEL.md](/BRANCHING_MODEL.md) file for best practices on how to use branches for this project.
+
+- Please refer to the [ROS C++ styling guide](http://wiki.ros.org/CppStyleGuide) that is based on [Google C++ style guide](https://google.github.io/styleguide/cppguide.html) as much as possible to keep the code readable!
+
+## Do you have questions?
+
+- Are they about how to use the tools? Refer to the relevant documentation. 
+
+- You are following the documentation and something still seems broken? Discuss on the discourse page (link to be added).
+
+- You are a contributor and want to contact the team about source code? Email at [info@cognicept.systems](mailto:info@cognicept.systems). Note that, support questions emailed won’t be addressed. Forums are the best place for that! 
+
+
+Thanks!
