+++ get_code_cov.bash
@@ -3,7 +3,7 @@ cd ~/ros2_dd_ws
 # Testing the robot 
 colcon build --symlink-install --cmake-args -DCMAKE_BUILD_TYPE=Debug -DCMAKE_CXX_FLAGS="--coverage" -DCMAKE_CXX_OUTPUT_EXTENSION_REPLACE=ON
 # Testing for code coverage
-colcon test --packages-select rosrect-listener-agent-ros2
+colcon test --packages-select error_resolution_diagnoser_ros2
 # LCOV info file generated
 # lcov --remove 'test/*' 'include/*' --capture --directory build/ --output-file ros2.info
 # GCOV report
