+++ include/error_resolution_diagnoser_ros2/robot_event.h
