--- launch/listener-agent-launch.py
@@ -1,11 +1,0 @@
-"""Launch a rosrect-listener-agent-ros2"""
-
-from launch import LaunchDescription
-import launch_ros.actions
-
-
-def generate_launch_description():
-    return LaunchDescription([
-        launch_ros.actions.Node(
-            package='rosrect-listener-agent-ros2', node_executable='rosrect-listener-agent-ros2', output='screen', arguments=['__log_disable_rosout:=true']),
-    ])
