+++ package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>rosrect-listener-agent-ros2</name>
+  <name>error_resolution_diagnoser_ros2</name>
   <version>0.0.0</version>
-  <description>The rosrect-listener-agent-ros2 package</description>
+  <description>The error_resolution_diagnoser_ros2 package</description>
   <maintainer email="swarooph@cognicept.systems">swaroophs</maintainer>
   <license>BSD</license>
