+++ startup_docker_container.bash
@@ -2,7 +2,7 @@
 
 clear
 
-docker build -t rosrect_agent_ros2 .
+docker build -t error_resolution_diagnoser_ros2 .
 
 docker stop agent
 
@@ -13,7 +13,5 @@ docker run -it \
 --network=host \
 --name=agent  \
 --volume="${HOME}/.cognicept/agent/logs:/root/.cognicept/agent/logs" \
-rosrect_agent_ros2:latest  \
-ros2 launch rosrect-listener-agent-ros2 listener-agent-launch.py
-
-# --env="ROS_MASTER_URI=http://localhost:11311" \
+error_resolution_diagnoser_ros2:latest  \
+ros2 launch error_resolution_diagnoser_ros2 error_resolution_diagnoser_ros2_launch.py
