+++ test/listener_integration_launch_db_test.py
@@ -24,8 +24,8 @@ def generate_test_description(ready_fn):
     proc_env['PYTHONUNBUFFERED'] = '1'
 
     listener_node = launch_ros.actions.Node(
-        package='rosrect-listener-agent-ros2',
-        node_executable='rosrect-listener-agent-ros2',
+        package='error_resolution_diagnoser_ros2',
+        node_executable='error_resolution_diagnoser_ros2',
         arguments=['__log_disable_rosout:=true'],
         env=proc_env,
     )
