+++ test/utests_robotevent.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 #include <fstream>
-#include <rosrect-listener-agent-ros2/robot_event.h>
+#include <error_resolution_diagnoser_ros2/robot_event.h>
 
 using namespace web::json; // JSON features
 using namespace web;       // Common features like URIs.
