+++ test/utests_statemanager.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
 #include <fstream>
 #include <boost/filesystem.hpp>
-#include <rosrect-listener-agent-ros2/state_manager.h>
+#include <error_resolution_diagnoser_ros2/state_manager.h>
 
 using namespace web::json; // JSON features
 using namespace web;       // Common features like URIs.
